🐞 Fix(hub): my private architectures 페이지에서 useSearchParams suspense 문제 해결
